pyproject.toml
[tool.poetry]
name = "drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.3.0"
django = "^5.0.6"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.1"
django-phonenumber-field = "^7.3.0"
phonenumbers = "^8.13.36"
python-dotenv = "^1.0.1"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
